call-service/requirements.txt
flask
twilio
SpeechRecognition==2.1.3
openai
elevenlabs
pydub
python-dotenv